demos/voice/requirements.txt
openai-whisper
piper-tts
sounddevice
numpy
